moved myEventsPage to client rendering
