--- src/app/[locale]/events/page.module.css
@@ -1,19 +1,0 @@
-.container-events {
-    display: flex;
-    width: 100%;
-    height: auto;
-    flex-direction: column;
-    align-items: center;
-    padding: var(--padding);
-    color: var(--color-whywhite-500);
-
-    .events {
-        width: 100%;
-        max-width: 700px;
-
-        display: flex;
-        flex-direction: column;
-        margin: 0 auto;
-        gap: 16px;
-    }
-}
