+++ src/app/[locale]/events/page.tsx
@@ -1,17 +1,8 @@
 import { getTranslator } from 'next-intl/server'
 
-import { useTranslations } from 'next-intl'
-
-import { useSession } from 'next-auth/react'
-
-import Title from '../components/atoms/ui/Title'
-
-import styles from './page.module.css'
+import { MyEvents } from '@/components/organisms/events/my-events/MyEvents'
 
 import { MetadataProps } from '@/app/layout.types'
-
-import { UserEvents } from '@/components/user-events/UserEvents'
-import { getEventsByUserName } from '@/services/event'
 
 export const generateMetadata = async ({
   params: { locale },
@@ -23,47 +14,10 @@ export const generateMetadata = async ({
   }
 }
 
-const controller = new AbortController()
-const signal = controller.signal
-controller.abort()
-
 const EventsPage = async () => {
-  // const t = useTranslations('Events')
-  // const { data: session } = useSession()
-  const body = {
-    // userName: session?.user?.name as string,
-    userName: 'ejrb1234',
-  }
-  const response = await getEventsByUserName(body)
-  const { data: events } = response
-  // console.log(response)
-
-  /*
-  TODO:
-  - used 'any' in typeScript -> inside .map
-  */
-
   return (
     <>
-      <div className={styles['container-events']}>
-        <Title>{'Mis eventos'}</Title>
-        {/* <Title>{t('title')}</Title> */} {/* error de async */}
-        <main className={styles['events']}>
-          {/* <p>{JSON.stringify(events)}</p> */}
-          {events?.map((event: any) => {
-            return (
-              <UserEvents
-                key={event.id}
-                date={event.createdAt}
-                description={event.description}
-                name={event.name}
-                timeZone={event.timeZone}
-                url={event.url}
-              />
-            )
-          })}
-        </main>
-      </div>
+      <MyEvents />
     </>
   )
 }
